backend/requirements.txt
fastapi
uvicorn
pyserial
google-generativeai
python-dotenv
scikit-learn
joblib
pandas
pyttsx3
edge-tts
pygame
